+++ packages/cds-annotation-parser/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.1
 
+*Released: 2026-06-03T14:58:37Z*
+
 ### Patch Changes
 
 - 21a3de7: FIX: TypeScript type errors in test files (ESM migration follow-up)
@@ -12,6 +14,8 @@
     - @sap-ux/odata-vocabularies@1.0.1
 
 ## 1.0.0
+
+*Released: 2026-05-30T20:54:07Z*
 
 ### Major Changes
 
@@ -54,6 +58,8 @@
 
 ## 0.3.1
 
+*Released: 2026-05-20T13:39:22Z*
+
 ### Patch Changes
 
 - @sap-ux/odata-annotation-core@0.3.1
@@ -61,6 +67,8 @@
 - @sap-ux/odata-vocabularies@0.5.1
 
 ## 0.3.0
+
+*Released: 2026-05-15T08:12:20Z*
 
 ### Minor Changes
 
@@ -76,6 +84,8 @@
 
 ## 0.2.37
 
+*Released: 2026-05-14T11:45:51Z*
+
 ### Patch Changes
 
 - 50a8ba5: chore: fresh release after workflow updates
@@ -87,6 +97,8 @@
 
 ## 0.2.36
 
+*Released: 2026-04-30T14:23:24Z*
+
 ### Patch Changes
 
 - Updated dependencies [c160401]
@@ -96,6 +108,8 @@
     - @sap-ux/odata-vocabularies@0.4.31
 
 ## 0.2.35
+
+*Released: 2026-03-30T22:24:11Z*
 
 ### Patch Changes