+++ packages/deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/deploy-config-sub-generator
 
+## 1.0.21
+
+### Patch Changes
+
+#### Release Date
+
+2026-06-18
+
+#### Workspace Updates
+
+- @sap-ux/cf-deploy-config-sub-generator 1.0.14 → 1.0.15
+- @sap-ux/abap-deploy-config-sub-generator 1.0.18 → 1.0.19
+
 ## 1.0.20
 
 ### Patch Changes
@@ -50,6 +63,8 @@
 
 ## 1.0.16
 
+*Released: 2026-06-14T10:40:09Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@1.0.16
@@ -57,12 +72,16 @@
 
 ## 1.0.15
 
+*Released: 2026-06-12T10:49:08Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@1.0.15
 - @sap-ux/odata-service-inquirer@3.0.11
 
 ## 1.0.14
+
+*Released: 2026-06-12T06:53:23Z*
 
 ### Patch Changes
 
@@ -75,12 +94,16 @@
 
 ## 1.0.13
 
+*Released: 2026-06-11T13:37:16Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@1.0.13
 - @sap-ux/odata-service-inquirer@3.0.10
 
 ## 1.0.12
+
+*Released: 2026-06-11T10:54:17Z*
 
 ### Patch Changes
 
@@ -92,6 +115,8 @@
 - @sap-ux/deploy-config-generator-shared@1.0.10
 
 ## 1.0.11
+
+*Released: 2026-06-10T09:57:42Z*
 
 ### Patch Changes
 
@@ -107,6 +132,8 @@
 
 ## 1.0.10
 
+*Released: 2026-06-09T14:35:01Z*
+
 ### Patch Changes
 
 - Updated dependencies [0fa8305]
@@ -120,6 +147,8 @@
 
 ## 1.0.9
 
+*Released: 2026-06-09T13:18:16Z*
+
 ### Patch Changes
 
 - Updated dependencies [a328e14]
@@ -131,6 +160,8 @@
     - @sap-ux/odata-service-inquirer@3.0.7
 
 ## 1.0.8
+
+*Released: 2026-06-04T13:54:21Z*
 
 ### Patch Changes
 
@@ -145,12 +176,16 @@
 
 ## 1.0.7
 
+*Released: 2026-06-04T12:10:05Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@1.0.7
 - @sap-ux/odata-service-inquirer@3.0.5
 
 ## 1.0.6
+
+*Released: 2026-06-04T10:19:37Z*
 
 ### Patch Changes
 
@@ -165,6 +200,8 @@
     - @sap-ux/deploy-config-generator-shared@1.0.5
 
 ## 1.0.5
+
+*Released: 2026-06-03T14:58:37Z*
 
 ### Patch Changes
 
@@ -183,6 +220,8 @@
 
 ## 1.0.4
 
+*Released: 2026-06-03T13:52:44Z*
+
 ### Patch Changes
 
 - Updated dependencies [9580241]
@@ -197,6 +236,8 @@
 
 ## 1.0.3
 
+*Released: 2026-06-02T11:35:17Z*
+
 ### Patch Changes
 
 - @sap-ux/deploy-config-generator-shared@1.0.2
@@ -208,12 +249,16 @@
 
 ## 1.0.2
 
+*Released: 2026-06-02T08:56:31Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@1.0.2
 - @sap-ux/odata-service-inquirer@3.0.1
 
 ## 1.0.1
+
+*Released: 2026-06-01T15:15:26Z*
 
 ### Patch Changes
 
@@ -227,6 +272,8 @@
     - @sap-ux/deploy-config-generator-shared@1.0.1
 
 ## 1.0.0
+
+*Released: 2026-05-30T20:54:07Z*
 
 ### Major Changes
 
@@ -275,6 +322,8 @@
 
 ## 0.6.12
 
+*Released: 2026-05-29T12:50:34Z*
+
 ### Patch Changes
 
 - Updated dependencies [a1c9cd7]
@@ -286,12 +335,16 @@
 
 ## 0.6.11
 
+*Released: 2026-05-29T06:59:27Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.4.10
 - @sap-ux/odata-service-inquirer@2.23.3
 
 ## 0.6.10
+
+*Released: 2026-05-27T11:39:21Z*
 
 ### Patch Changes
 
@@ -303,6 +356,8 @@
 - @sap-ux/inquirer-common@0.13.5
 
 ## 0.6.9
+
+*Released: 2026-05-27T10:42:47Z*
 
 ### Patch Changes
 
@@ -318,6 +373,8 @@
 
 ## 0.6.8
 
+*Released: 2026-05-26T16:40:21Z*
+
 ### Patch Changes
 
 - Updated dependencies [01b70ca]
@@ -332,6 +389,8 @@
 
 ## 0.6.7
 
+*Released: 2026-05-25T14:44:33Z*
+
 ### Patch Changes
 
 - Updated dependencies [acb9875]
@@ -339,12 +398,16 @@
 
 ## 0.6.6
 
+*Released: 2026-05-22T13:30:05Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.4.6
 - @sap-ux/odata-service-inquirer@2.22.2
 
 ## 0.6.5
+
+*Released: 2026-05-21T16:21:11Z*
 
 ### Patch Changes
 
@@ -360,12 +423,16 @@
 
 ## 0.6.4
 
+*Released: 2026-05-21T14:58:44Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.4.4
 - @sap-ux/odata-service-inquirer@2.22.1
 
 ## 0.6.3
+
+*Released: 2026-05-19T15:16:46Z*
 
 ### Patch Changes
 
@@ -380,12 +447,16 @@
 
 ## 0.6.2
 
+*Released: 2026-05-18T08:15:14Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.4.2
 - @sap-ux/odata-service-inquirer@2.22.0
 
 ## 0.6.1
+
+*Released: 2026-05-15T13:12:06Z*
 
 ### Patch Changes
 
@@ -402,6 +473,8 @@
     - @sap-ux/deploy-config-generator-shared@0.2.1
 
 ## 0.6.0
+
+*Released: 2026-05-15T08:12:20Z*
 
 ### Minor Changes
 
@@ -423,6 +496,8 @@
 
 ## 0.5.161
 
+*Released: 2026-05-14T21:28:41Z*
+
 ### Patch Changes
 
 - @sap-ux/inquirer-common@0.11.48
@@ -433,11 +508,15 @@
 
 ## 0.5.160
 
+*Released: 2026-05-14T14:16:50Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.36
 
 ## 0.5.159
+
+*Released: 2026-05-14T11:45:51Z*
 
 ### Patch Changes
 
@@ -456,6 +535,8 @@
 
 ## 0.5.158
 
+*Released: 2026-05-13T09:36:59Z*
+
 ### Patch Changes
 
 - Updated dependencies [21abda3]
@@ -469,6 +550,8 @@
 
 ## 0.5.157
 
+*Released: 2026-05-12T18:00:39Z*
+
 ### Patch Changes
 
 - Updated dependencies [9360ea5]
@@ -480,6 +563,8 @@
     - @sap-ux/odata-service-inquirer@2.20.22
 
 ## 0.5.156
+
+*Released: 2026-05-06T23:02:00Z*
 
 ### Patch Changes
 
@@ -497,6 +582,8 @@
 
 ## 0.5.155
 
+*Released: 2026-05-01T15:46:09Z*
+
 ### Patch Changes
 
 - @sap-ux/cf-deploy-config-sub-generator@0.2.178
@@ -504,11 +591,15 @@
 
 ## 0.5.154
 
+*Released: 2026-04-30T19:47:20Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.30
 
 ## 0.5.153
+
+*Released: 2026-04-30T14:23:24Z*
 
 ### Patch Changes
 
@@ -522,6 +613,8 @@
 
 ## 0.5.152
 
+*Released: 2026-04-30T13:10:33Z*
+
 ### Patch Changes
 
 - @sap-ux/odata-service-inquirer@2.20.19
@@ -529,12 +622,16 @@
 
 ## 0.5.151
 
+*Released: 2026-04-29T16:22:20Z*
+
 ### Patch Changes
 
 - Updated dependencies [8b2c55b]
     - @sap-ux/odata-service-inquirer@2.20.18
 
 ## 0.5.150
+
+*Released: 2026-04-29T15:24:37Z*
 
 ### Patch Changes
 
@@ -549,6 +646,8 @@
 
 ## 0.5.149
 
+*Released: 2026-04-27T19:47:46Z*
+
 ### Patch Changes
 
 - Updated dependencies [1d60871]
@@ -562,11 +661,15 @@
 
 ## 0.5.148
 
+*Released: 2026-04-27T07:30:24Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.25
 
 ## 0.5.147
+
+*Released: 2026-04-23T12:54:21Z*
 
 ### Patch Changes
 
@@ -581,6 +684,8 @@
 
 ## 0.5.146
 
+*Released: 2026-04-23T06:48:55Z*
+
 ### Patch Changes
 
 - Updated dependencies [237371b]
@@ -593,6 +698,8 @@
 
 ## 0.5.145
 
+*Released: 2026-04-22T12:38:46Z*
+
 ### Patch Changes
 
 - @sap-ux/inquirer-common@0.11.38
@@ -603,11 +710,15 @@
 
 ## 0.5.144
 
+*Released: 2026-04-20T12:41:53Z*
+
 ### Patch Changes
 
 - @sap-ux/cf-deploy-config-sub-generator@0.2.171
 
 ## 0.5.143
+
+*Released: 2026-04-15T11:53:17Z*
 
 ### Patch Changes
 
@@ -621,11 +732,15 @@
 
 ## 0.5.142
 
+*Released: 2026-04-15T08:11:32Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.20
 
 ## 0.5.141
+
+*Released: 2026-04-14T20:26:28Z*
 
 ### Patch Changes
 
@@ -639,6 +754,8 @@
     - @sap-ux/odata-service-inquirer@2.20.11
 
 ## 0.5.140
+
+*Released: 2026-04-14T12:35:35Z*
 
 ### Patch Changes
 
@@ -655,11 +772,15 @@
 
 ## 0.5.139
 
+*Released: 2026-04-14T11:39:16Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.17
 
 ## 0.5.138
+
+*Released: 2026-04-13T13:57:02Z*
 
 ### Patch Changes
 
@@ -667,17 +788,23 @@
 
 ## 0.5.137
 
+*Released: 2026-04-09T09:57:18Z*
+
 ### Patch Changes
 
 - @sap-ux/cf-deploy-config-sub-generator@0.2.166
 
 ## 0.5.136
 
+*Released: 2026-04-09T07:00:16Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.16
 
 ## 0.5.135
+
+*Released: 2026-04-08T13:10:18Z*
 
 ### Patch Changes
 
@@ -695,11 +822,15 @@
 
 ## 0.5.134
 
+*Released: 2026-04-07T11:09:34Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.14
 
 ## 0.5.133
+
+*Released: 2026-04-02T13:21:06Z*
 
 ### Patch Changes
 
@@ -707,17 +838,23 @@
 
 ## 0.5.132
 
+*Released: 2026-04-02T12:50:13Z*
+
 ### Patch Changes
 
 - @sap-ux/cf-deploy-config-sub-generator@0.2.163
 
 ## 0.5.131
 
+*Released: 2026-04-01T14:51:40Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.13
 
 ## 0.5.130
+
+*Released: 2026-04-01T11:49:37Z*
 
 ### Patch Changes
 
@@ -732,11 +869,15 @@
 
 ## 0.5.129
 
+*Released: 2026-04-01T06:34:51Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.11
 
 ## 0.5.128
+
+*Released: 2026-03-30T22:24:11Z*
 
 ### Patch Changes
 
@@ -761,6 +902,8 @@
 
 ## 0.5.127
 
+*Released: 2026-03-27T15:37:24Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.9
@@ -771,6 +914,8 @@
 - @sap-ux/deploy-config-generator-shared@0.1.112
 
 ## 0.5.126
+
+*Released: 2026-03-27T11:58:49Z*
 
 ### Patch Changes
 
@@ -785,6 +930,8 @@
 
 ## 0.5.125
 
+*Released: 2026-03-26T20:06:10Z*
+
 ### Patch Changes
 
 - Updated dependencies [b66e827]
@@ -798,17 +945,23 @@
 
 ## 0.5.124
 
+*Released: 2026-03-26T16:48:32Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.6
 
 ## 0.5.123
 
+*Released: 2026-03-26T15:15:10Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.5
 
 ## 0.5.122
+
+*Released: 2026-03-26T12:07:04Z*
 
 ### Patch Changes
 
@@ -836,6 +989,8 @@
 
 ## 0.5.121
 
+*Released: 2026-03-25T12:56:41Z*
+
 ### Patch Changes
 
 - Updated dependencies [f384ace]
@@ -849,6 +1004,8 @@
 
 ## 0.5.120
 
+*Released: 2026-03-23T18:25:40Z*
+
 ### Patch Changes
 
 - @sap-ux/abap-deploy-config-sub-generator@0.3.2
@@ -858,11 +1015,15 @@
 
 ## 0.5.119
 
+*Released: 2026-03-23T14:11:51Z*
+
 ### Patch Changes
 
 - @sap-ux/cf-deploy-config-sub-generator@0.2.155
 
 ## 0.5.118
+
+*Released: 2026-03-23T10:51:10Z*
 
 ### Patch Changes