+++ packages/i18n/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.0.1
 
+*Released: 2026-06-03T14:58:37Z*
+
 ### Patch Changes
 
 - 21a3de7: FIX: TypeScript type errors in test files (ESM migration follow-up)
@@ -9,6 +11,8 @@
     - @sap-ux/text-document-utils@1.0.1
 
 ## 1.0.0
+
+*Released: 2026-05-30T20:54:07Z*
 
 ### Major Changes
 
@@ -48,6 +52,8 @@
 
 ## 0.4.0
 
+*Released: 2026-05-15T08:12:20Z*
+
 ### Minor Changes
 
 - 72695e5: chore: drop Node 20 support as it is no longer maintained
@@ -59,6 +65,8 @@
 
 ## 0.3.12
 
+*Released: 2026-05-14T11:45:51Z*
+
 ### Patch Changes
 
 - 50a8ba5: chore: fresh release after workflow updates
@@ -67,12 +75,16 @@
 
 ## 0.3.11
 
+*Released: 2026-04-30T14:23:24Z*
+
 ### Patch Changes
 
 - Updated dependencies [c160401]
     - @sap-ux/text-document-utils@0.3.4
 
 ## 0.3.10
+
+*Released: 2026-03-26T12:07:04Z*
 
 ### Patch Changes